Create a django command for importing xml codes
Correos de Mexico offers a xml file that contains all the postal codes,
and their respective details, from Mexico. However this file should be
processed before it can be uploaded to the database. The purporse of
creating a django command is to keep the database separated from the
code implementation.