Refactor code from importantpostalcodes command
Instead of hardcoding the required columns to create the database, the
code now states those columns in a list and uses a dict comprehension
for creating dynamically the necessary kwargs to instantiate a
PostalCode object. Also, since all the columns start with '{NewDataSet}'
string, this commit stores this string as a variable and uses it as a
prefix before generating the column names dynamically.