Enable manual file import and refactor command
Add --file argument to command's parser if this argument is found the
code will attempt to obtain the xml data from the file specified at
the same level as the manage.py file.
if no file name is specified the default one will be used (CPdescarga.xml)
This commit also modifies lightly the error messages.