Prevent command to create database twice
The importpostalcodemx command now check whether a postal code exist or
not, if it exists, it exits without doing anything, otherwise it
continues and create the required tables. The purpose of this is to
create the tables only once, multiple runs of the command create
database inconsistencies and duplication of fields, because bulk_create
method, used to speed up the table creation, ignores conflicts.